Webhook tests grouped and refactored
